fix: align RevenueCat workspace lockfile to 5.68.0 (#239)
* chore(deps): bump github.com/revenuecat/purchases-ios

Bumps [github.com/revenuecat/purchases-ios](https://github.com/RevenueCat/purchases-ios) from 4.44.2 to 5.68.0.
- [Release notes](https://github.com/RevenueCat/purchases-ios/releases)
- [Commits](https://github.com/RevenueCat/purchases-ios/compare/4.44.2...5.68.0)

---
updated-dependencies:
- dependency-name: github.com/revenuecat/purchases-ios
  dependency-version: 5.68.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* fix: align RevenueCat workspace lockfile

* fix: restore iOS lint compliance

---------
