+++ ios/OpenClawConsole/Tests/GitViewModelTests.swift
